[ADD] point_of_sale: rescue sessions for orders on closed sessions
When an order is linked to a closed session, create a new rescue session for
these orphan orders. Orders should not be added to closed session as accounting
entries are already generated.

Add boolean field rescue (readonly and hidden to make sure user can not create
rescue sessions).
Ignore the unicity constraints for the rescue sessions.