[FIX] account_followup: partial overdue payments
This is related to rev. 7402e898204f1a1f492d56848780e5ae92c5297f

The above rev. take care to exclude payments that are not yet due;
meaning the ones due in the future, by checking the maturity
date.

Problem: Payments (e.g. move lines from bank statements)
do not have a maturity date. Only move lines that actually
have a maturity date, in the future, must be excluded,
not the one that do not have a maturity date.

opw-633930