[REV] website_hr_recruitment: revert 1e895ab
This commit would makes us very vulnerable to spams and malevolent
bots who would forge e-mails to force the creation of a res_partner
for each e-mail, quickly overloading the database. I'm reverting
this as it was requested by al.

After discussion with yti, the original issue was that to use the
mail thread on the hr application, the user needed to create a
res_partner just to send him a mail, even if he was refused.

A proposed solution would be to add the e-mail of the applicant
to the suggested recipients of the mail thread.