[FIX] account: Exchanges rates gain/loss during supplier reconciliations
This revision is related to 3d2b38844622fde6816edc20362bf35bc496c5e6.

The above revision was allmost correct, except that, in a supplier
invoices, the `amount_currency` of the invoice is negative,
and the one of the bank statement is possitive.

To check that both are equal, we should subtract one to each other,
and check that the diff is 0, but both part need to be the
absolute value.

opw-634263
closes #6533