[FIX] product: supplier pricelist field in partner form
This is related to revision be0687df54884cbc9e012eab382d9a540682cf16.

The inherited view adding
the purchase pricelist field on the partner form
no longer worked for user being in the purchase pricelist group
but not in the sale pricelist group, since the field was added
after the sale pricelist field, which was no longer in the view
for these users.

This revision partially revert the above revision, but
add the purchase pricelist group in the groups of the view,
so the view is loaded only if the user is part of one group
or the other.

opw-639685