[FIX] pos_restaurant: changed unit caused issue on splitting
Commit 13af42d83 changed the unit rounding from units from 1 to 0.001,
it was done mainly because it's the default unit and in the interface we
see quantities with 3 decimals after the floating point, so people might
expect to be able to set floating point quantities.

This change caused an issue when splitting a bill in the pos restaurant,
instead of a click being a piece of the product, it was 0.001 part of a
product.

This fix overrides the rounding of units in this case.

closes #6780
opw-639865