[FIX] report: datetime already imported, differently
This rev. is related to 5f93103361ba424d1d834a828685030c53b9a3ab.

The module datetime was already available in the reports,
and it was the entire datetime module that was imported,
not just the class.

This rev. is not recompatible with 5f93103361ba424d1d834a828685030c53b9a3ab,
but there is no other choice, since other existing
reports might already expected the datetime module
and not the class

opw-640299