[FIX] stock_account: Create invoice journal type for transit.
This revision is related to 29ebf0b4b1adf45337c8d012d908f0612273bb17.

For a 'To be invoiced' delivery order with a move with
as source location 'Stock' and
as destination location 'Transit'

There is no way to know if it's a Customer invoice
that should be created or a Purchase Invoice.
Indeed, the transit could be used as an intermediate to ship
to the supplier, or to the customer

opw-639536