[FIX] account: default values of payments (customer/supplier)
In `account.abstract.payment` model,
The `partner_type` and `payment_type` fields had set as default
`supplier` and `outbound`, respectively.

The thing is, the `default_get` method of model `account.payment`
(inheriting on the above model), will keep these default values set,
even if the invoice is customer invoice, which is wrong.

Either, this `default_get` should ignore the current default values,
either, the default values should not be wrong in the first place.

In this revision, I choose the second solution. Therefore, I had
to add the default value `outbound` in the purchase payments
menu action.

opw-641441