Revert "[FIX] web: default loaded view on direct access/refresh"
This reverts commit ee9a4d4f8a5e974153b672f21e3ac0c77b43be75.

By putting into the flags object the default_view stored into the
context (i.e. the one written in the url), it happens that a View
Manager tries to open a view which is not available. For instance,
Sales > Sales Teams. Then refresh the page. As we are in a Kanban
view, the view_type = Kanban is stored in the context. Then click
on Quotations in a Kanban card. It doesn't work because the
ViewManager tries to open the Kanban View of the Quotations which
doesn't exist.