[FIX] google_calendar: fail during first sync
If an event exists on Google but was not synced yet in Odoo, the synchronisation
would fail.
When self.OE.found is False, self.OE.event is not defined and trying to access
it would fail (it's a python object, not a recordset)

Check the ownership after making sure the event is present in Odoo.

Fixes #7034