[FIX] hr_gamification: user_id not required
Revert the modification from commit b2ef4969e9eac9aeb232c89f1c0d7413c17a400c ; it appears that the test of required field is done before the computation of the related field and that leads to a Validation Error.

As the field employee_id is already required and the user_id is related to this field, the attribute required is redundant and unecessary.