[IMP] base_action_rule: allow to use time in domain
Unlink typical domain evaluation (ir.rule, filters,...), the evaluated domain
for automated actions did not included time in the context so it was not
possible to make time based conditions in domain.

This should be used with care as filters 'Based on Timed Condition' are still
possible and will probably be enough (and safer) in most cases.